Upgraded celo-blockchain deployed version to 1.5.5 (#9357)
### Description

Bumped the celo-blockchain version in production environments to [1.5.5](https://github.com/celo-org/celo-blockchain/releases/tag/v1.5.5).

Already deployed to alfajores and baklava.

+++ packages/celotool/src/lib/helm_deploy.ts
@@ -444,9 +444,13 @@ export async function registerIPAddress(name: string, zone?: string) {
 export async function deleteIPAddress(name: string, zone?: string) {
   console.info(`Deleting IP address ${name}`)
   try {
-    await execCmd(
-      `gcloud compute addresses delete ${name} --region ${getKubernetesClusterRegion(zone)} -q`
-    )
+    if (isCelotoolVerbose()) {
+      console.info(`IP Address ${name} would be deleted`)
+    } else {
+      await execCmd(
+        `gcloud compute addresses delete ${name} --region ${getKubernetesClusterRegion(zone)} -q`
+      )
+    }
   } catch (error: any) {
     if (!error.toString().includes('was not found')) {
       console.error(error)
